Adicionar o Get todas as contas

+++ Application/Arguments/Conta/Adicionar/AdicionarContaRequest.cs
@@ -5,7 +5,7 @@ namespace Application.Arguments.Conta.Adicionar
     public class AdicionarContaRequest 
     {
         public string Nome { get; set; }
-        public long Valor { get; set; }
+        public double Valor { get; set; }
         public DateTime? DataVencimento { get; set; }
 
     }
