+++ Domain/Entities/Conta.cs
@@ -8,7 +8,7 @@ namespace Domain.Entities
     {
         public string Nome { get; set; }
         public long ValorOriginal { get; set; }
-        public  long? ValorCorrigido { get; set; }
+        public  double? ValorCorrigido { get; set; }
         public DateTime  DataVencimento  { get; set; }
         public DateTime DataPagamento { get; set; }
         public StatusEnum status { get; set; }
